Create whole.mli
interface file that restricts what can be seen and what cannot be seen. 
Everything that is included here will be visible to other .ml file that tries to use contents in whole.ml file. This file will restrict access to any element in whole.ml file if they are not listed in this file