Create whole_use.ml
this will try to use elements in whole.ml file.
line 11 is commented out because it will try to access the element that should not be accessible. you_cant_see_me variable is not listed inside the whole.mli file. So if you compile it what line 11 and 12 uncommeted, it will not compile at all.